// src/store/roleAndPermission/rolesAndPermissions.enum.ts
enum RolesAndPermissionTypeEnum {
  GET_ROLES_LOADING = "GET_ROLES_LOADING",
  GET_ROLES_LOADED = "GET_ROLES_LOADED",
  GET_ROLES_FAILED = "GET_ROLES_FAILED",

  GET_MODULES_LOADING = "GET_MODULES_LOADING",
  GET_MODULES_LOADED = "GET_MODULES_LOADED",
  GET_MODULES_FAILED = "GET_MODULES_FAILED",

  GET_ROLES = "GET_ROLES",
  GET_ROLES_BY_ID = "GET_ROLES_BY_ID",
  ADD_ROLES = "ADD_ROLES",
  UPDATE_ROLES_AND_PERMISSION = "UPDATE_ROLES_AND_PERMISSION",
  DELETE_ROLES = "DELETE_ROLES",
  UPDATE_ROLES_STATUS = "UPDATE_ROLES_STATUS",

  GET_MODULES = "GET_MODULES",
  
}
export default RolesAndPermissionTypeEnum;